changes on e2e-0 to test/e2e/app_autosync/app_autosync_test.go,test/e2e/app_deletion/app_deletion_test.go,test/e2e/app_management/app_management_test.go,test/e2e/app_of_apps/app_of_apps_test.go,test/e2e/cli/cli_test.go,test/e2e/cluster/cluster_test.go,test/e2e/cluster_objects/cluster_objects_test.go,test/e2e/custom_tool/custom_tool_test.go,test/e2e/delarative/delarative_test.go,test/e2e/deployment/deployment_test.go,test/e2e/diff/diff_test.go,test/e2e/fixture/app/expectation.go,test/e2e/fixture/certs/certs.go,test/e2e/fixture/fixture.go,test/e2e/fixture/repos/repos.go,test/e2e/git_submodule/git_submodule_test.go,test/e2e/helm/helm_test.go,test/e2e/hook/hook_test.go,test/e2e/jsonnet/jsonnet_test.go,test/e2e/kustomize/kustomize_test.go,test/e2e/project_management/project_management_test.go,test/e2e/pruning_required/pruning_required_test.go,test/e2e/repo_creds/repo_creds_test.go,test/e2e/repo_management/repo_management_test.go,test/e2e/selective_sync/selective_sync_test.go,test/e2e/ssh_repo/ssh_repo_test.go,test/e2e/sync_options/sync_options_test.go,test/e2e/sync_waves/sync_waves_test.go,test/e2e/user_info/user_info_test.go,
