--- test/e2e/cluster_objects_test.go
@@ -1,23 +1,0 @@
-//+build group1
-
-package e2e
-
-import (
-	"testing"
-
-	. "github.com/argoproj/argo-cd/pkg/apis/application/v1alpha1"
-	. "github.com/argoproj/argo-cd/test/e2e/fixture/app"
-)
-
-// ensure that cluster scoped objects, like a cluster role, as a hok, can be successfully deployed
-func TestClusterRoleBinding(t *testing.T) {
-	Given(t).
-		Path("cluster-role").
-		When().
-		Create().
-		Sync().
-		Then().
-		Expect(OperationPhaseIs(OperationSucceeded)).
-		Expect(HealthIs(HealthStatusHealthy)).
-		Expect(SyncStatusIs(SyncStatusCodeSynced))
-}
