+++ test/e2e/fixture/certs/certs.go
@@ -11,7 +11,7 @@ import (
 // Add a custom CA certificate to the test and also create the certificate file
 // on the file system, so argocd-server and argocd-repo-server can use it.
 func AddCustomCACert() {
-	caCertPath, err := filepath.Abs("../fixture/certs/argocd-test-ca.crt")
+	caCertPath, err := filepath.Abs("../../fixture/certs/argocd-test-ca.crt")
 	errors.CheckError(err)
 	args := []string{"cert", "add-tls", "localhost", "--from", caCertPath}
 	errors.FailOnErr(fixture.RunCli(args...))
@@ -27,7 +27,7 @@ func AddCustomCACert() {
 }
 
 func AddCustomSSHKnownHostsKeys() {
-	knownHostsPath, err := filepath.Abs("../fixture/certs/ssh_known_hosts")
+	knownHostsPath, err := filepath.Abs("../../fixture/certs/ssh_known_hosts")
 	errors.CheckError(err)
 	args := []string{"cert", "add-ssh", "--batch", "--from", knownHostsPath}
 	errors.FailOnErr(fixture.RunCli(args...))
