--- test/e2e/ssh_repo_test.go
@@ -1,37 +1,0 @@
-//+build group3
-
-package e2e
-
-import (
-	"testing"
-
-	"github.com/argoproj/argo-cd/test/e2e/fixture"
-
-	. "github.com/argoproj/argo-cd/pkg/apis/application/v1alpha1"
-	. "github.com/argoproj/argo-cd/test/e2e/fixture/app"
-)
-
-func TestCanAccessInsecureSSHRepo(t *testing.T) {
-	Given(t).
-		SSHInsecureRepoURLAdded(true).
-		RepoURLType(fixture.RepoURLTypeSSH).
-		Path("config-map").
-		When().
-		Create().
-		Sync().
-		Then().
-		Expect(OperationPhaseIs(OperationSucceeded))
-}
-
-func TestCanAccessSSHRepo(t *testing.T) {
-	Given(t).
-		CustomSSHKnownHostsAdded().
-		SSHRepoURLAdded(true).
-		RepoURLType(fixture.RepoURLTypeSSH).
-		Path("config-map").
-		When().
-		Create().
-		Sync().
-		Then().
-		Expect(OperationPhaseIs(OperationSucceeded))
-}
