+++ test/e2e/user_info/user_info_test.go
@@ -1,0 +1,21 @@
+package user_info
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+
+	. "github.com/argoproj/argo-cd/test/e2e/fixture"
+)
+
+func TestUserInfo(t *testing.T) {
+	EnsureCleanState(t)
+
+	output, err := RunCli("account", "get-user-info")
+
+	assert.NoError(t, err)
+	assert.Equal(t, `Logged In: true
+Username: admin
+Issuer: argocd
+Groups: `, output)
+}
